Move server-mock objects into their own modules, so that we separate test code, from code that is meant to be used by users.
