+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/ComponentStatusTest.java
@@ -21,7 +21,7 @@ import io.fabric8.kubernetes.api.model.ComponentStatusBuilder;
 import io.fabric8.kubernetes.api.model.ComponentStatusList;
 import io.fabric8.kubernetes.api.model.ComponentStatusListBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.server.mock.KubernetesServer;
+import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 import org.junit.Rule;
 import org.junit.Test;
