+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/JobTest.java
@@ -22,7 +22,7 @@ import io.fabric8.kubernetes.api.model.JobList;
 import io.fabric8.kubernetes.api.model.JobListBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.kubernetes.server.mock.KubernetesServer;
+import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 import org.junit.Rule;
 import org.junit.Test;
