+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/SecurityContextConstraintsTest.java
@@ -21,7 +21,7 @@ import io.fabric8.kubernetes.api.model.SecurityContextConstraintsBuilder;
 import io.fabric8.kubernetes.api.model.SecurityContextConstraintsList;
 import io.fabric8.kubernetes.api.model.SecurityContextConstraintsListBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.server.mock.KubernetesServer;
+import io.fabric8.kubernetes.client.server.mock.KubernetesServer;
 import org.junit.Rule;
 import org.junit.Test;
