+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/mock/AdaptTest.java
@@ -19,7 +19,7 @@ package io.fabric8.openshift.client.mock;
 import okhttp3.OkHttpClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.openshift.client.OpenShiftClient;
-import io.fabric8.openshift.server.mock.OpenShiftMockServer;
+import io.fabric8.openshift.client.server.mock.OpenShiftMockServer;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
