+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/mock/BuildConfigTest.java
@@ -25,6 +25,8 @@ import io.fabric8.openshift.api.model.BuildConfigList;
 import io.fabric8.openshift.api.model.BuildConfigListBuilder;
 import io.fabric8.openshift.api.model.BuildListBuilder;
 import io.fabric8.openshift.client.OpenShiftClient;
+import io.fabric8.openshift.client.server.mock.OpenShiftServer;
+
 import org.junit.Rule;
 import org.junit.Test;
