+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/mock/DeploymentConfigTest.java
@@ -30,6 +30,7 @@ import io.fabric8.openshift.api.model.DeploymentConfigBuilder;
 import io.fabric8.openshift.api.model.DeploymentConfigList;
 import io.fabric8.openshift.api.model.DeploymentConfigListBuilder;
 import io.fabric8.openshift.client.OpenShiftClient;
+import io.fabric8.openshift.client.server.mock.OpenShiftServer;
 
 public class DeploymentConfigTest {
   @Rule
