+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/mock/LoadAsTemplateTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.client.mock;
+package io.fabric8.openshift.client.mock;
 
 import org.junit.Test;
