+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/mock/OAuthClientTest.java
@@ -21,6 +21,8 @@ import io.fabric8.openshift.api.model.OAuthClientBuilder;
 import io.fabric8.openshift.api.model.OAuthClientList;
 import io.fabric8.openshift.api.model.OAuthClientListBuilder;
 import io.fabric8.openshift.client.OpenShiftClient;
+import io.fabric8.openshift.client.server.mock.OpenShiftServer;
+
 import org.junit.Rule;
 import org.junit.Test;
