+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/mock/ProjectRequestTest.java
@@ -21,6 +21,8 @@ import io.fabric8.kubernetes.api.model.StatusBuilder;
 import io.fabric8.openshift.api.model.ProjectRequest;
 import io.fabric8.openshift.api.model.ProjectRequestBuilder;
 import io.fabric8.openshift.client.OpenShiftClient;
+import io.fabric8.openshift.client.server.mock.OpenShiftServer;
+
 import org.junit.Rule;
 import org.junit.Test;
